# docs/swagger.yaml
definitions:
...
  entities.Currency:
    properties:
      additional_alias_rate:
        type: number
      alias:
        type: string
      base_currency:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  requests.Currency:
    properties:
      additional_alias_rate:
        minimum: 1
        type: number
      alias:
        minLength: 1
        type: string
      base_currency:
        type: string
      title:
        minLength: 1
        type: string
      type:
        minLength: 1
        type: string
    required:
    - additional_alias_rate
    type: object
...
  response.Response:
    properties:
      data: {}
      meta: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/currency:
    delete:
      consumes:
      - application/json
      description: Delete currency.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: alias
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - X-Authenticate: []
      summary: Delete currency.
      tags:
      - currency
    get:
      consumes:
      - application/json
      description: Get all currencies` information.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Currency'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get all currencies.
      tags:
      - currency
    post:
      consumes:
      - application/json
      description: Create new currency.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CreateCurrency
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.Currency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Currency'
              type: object
      security:
      - X-Authenticate: []
      summary: Create new currency.
      tags:
      - currency
  /api/currency/{alias}:
    get:
      consumes:
      - application/json
      description: Get currency information.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: alias
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Currency'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get currency.
      tags:
      - currency
...
swagger: "2.0"